Support encryption via pre-shared key stored in RID file.
This is temporary. It is a place-holder while we move to a proper
key-exchange protocol where federates use public key crypto during
connection and exchange a session key that is then used at runtime.

Next steps are to get the basics of asym working and then implement
the exchange protocol.

Working: #240, #244